Agrego archivo de log (logueo algunas cosas aunque hay q revisarlas) y funcion volcarRamALog

+++ Memoria/src/configuracion.h
@@ -39,7 +39,7 @@ typedef struct{
 	int tamanioDeMarco;
 	int entradasDeTLB;
 	char* TLBHabilitada;
-	double retardoDeMemoria;
+	int retardoDeMemoria;
 	char* algoritmoRAM;
 }tipoConfigMemoria;
